app/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2==2.9.9
djangorestframework-simplejwt==5.3.1
# flake8>=3.6.0,<3.7.0
redis
celery
django-celery-results
SQLAlchemy